pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wm_tutorial"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "torch",
    "matplotlib",
    "rich",
    "descript-audiotools",
    "dasp-pytorch",
]

[tool.setuptools.packages.find]
where = ["."]